requirements.txt
# ============================================================================
# GraphRAG Vietnamese Law  —  Python Dependencies
# ============================================================================

# ---- Web Framework ----
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0

# ---- Database Drivers ----
psycopg2-binary==2.9.9
neo4j==5.25.0
redis==5.1.0

# ---- AI / ML  (Embedding + FAISS) ----
sentence-transformers==3.3.0
faiss-cpu==1.9.0
numpy==1.26.4
torch>=2.0.0

# ---- LLM Providers ----
openai>=1.40.0
anthropic>=0.34.0

# ---- LangChain  (text splitting + optional chains) ----
langchain>=0.3.0
langchain-community>=0.3.0

# ---- Utils ----
python-dotenv==1.0.1
